requirements.txt
torch
numpy
scipy
scikit-learn
matplotlib
scanpy
anndata
torchmetrics
pyyaml
